+++ lib/widgets/ny_form.dart
@@ -1,16 +1,23 @@
+import 'dart:math';
+
 import 'package:date_field/date_field.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:intl/intl.dart' as intl;
-import 'package:nylo_support/widgets/styles/bottom_modal_sheet_style.dart';
+import '/localization/app_localization.dart';
+import '/widgets/styles/bottom_modal_sheet_style.dart';
+import '/widgets/styles/ny_radio_tile_style.dart';
+import 'package:recase/recase.dart';
+
 import '/widgets/form/validation.dart';
 import '/widgets/ny_text_field.dart';
 import 'form/casts.dart';
-export 'form/form.dart';
+
 export 'form/casts.dart';
-export 'form/validation.dart';
+export 'form/form.dart';
 export 'form/form_data.dart';
+export 'form/validation.dart';
 
 /// [FormSubmitCallback] is a typedef that helps in managing form submit callbacks
 typedef FormSubmitCallback = (dynamic, Function(dynamic data))?;
@@ -272,6 +279,7 @@ extension NyFieldStyle on String {
 class Field {
   Field(
     this.key, {
+    this.label,
     this.value,
     FormCast? cast,
     this.validate,
@@ -303,6 +311,7 @@ class Field {
   /// Field.text is a constructor that helps in managing text fields
   Field.text(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -340,6 +349,7 @@ class Field {
   /// Field.currency is a constructor that helps in managing currency fields
   Field.currency(
     this.key, {
+    this.label,
     required String currency,
     this.value,
     this.validate,
@@ -371,6 +381,7 @@ class Field {
   /// Field.password is a constructor that helps in managing password fields
   Field.password(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -402,6 +413,7 @@ class Field {
   /// Field.email is a constructor that helps in managing password fields
   Field.email(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -439,6 +451,7 @@ class Field {
   /// Field.capitalizeWords is a constructor that helps in managing capitalizeWords fields
   Field.capitalizeWords(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -476,6 +489,7 @@ class Field {
   /// Field.capitalizeSentences is a constructor that helps in managing capitalizeSentences fields
   Field.capitalizeSentences(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -513,6 +527,7 @@ class Field {
   /// Field.picker is a constructor that helps in managing picker fields
   Field.picker(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -543,9 +558,57 @@ class Field {
     }
   }
 
+  /// Field.widget is a constructor that helps in managing widget fields
+  Field.widget({required Widget child})
+      : cast = FormCast.widget(
+          child: child,
+        ),
+        autofocus = false {
+    key = _randomKey();
+  }
+
+  /// Generate a random key
+  String _randomKey() {
+    return Random().nextInt(100000).toString();
+  }
+
+  /// Field.radio is a constructor that helps in managing radio fields
+  Field.radio(
+    this.key, {
+    this.label,
+    this.value,
+    this.validate,
+    this.autofocus = false,
+    this.dummyData,
+    this.header,
+    this.footer,
+    this.style,
+    this.metaData = const {},
+    this.hidden = false,
+    this.readOnly,
+    required List<String> options,
+    NyRadioTileStyle? nyRadioTileStyle,
+  }) : cast = FormCast.radio(
+            options: options, nyRadioTileStyle: nyRadioTileStyle) {
+    if (style == null) return;
+
+    metaData = {};
+    if (style is String) {
+      style = style;
+      return;
+    }
+    if (style is Map) {
+      style as Map<String, dynamic>;
+      metaData!["decoration_style"] =
+          (style as Map<String, dynamic>).entries.first.value;
+      style = (style as Map<String, dynamic>).entries.first.key;
+    }
+  }
+
   /// Field.number is a constructor that helps in managing number fields
   Field.number(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -577,6 +640,7 @@ class Field {
   /// Field.mask is a constructor that helps in managing mask fields
   Field.mask(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -618,6 +682,7 @@ class Field {
   /// Field.url is a constructor that helps in managing url fields
   Field.url(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -654,7 +719,8 @@ class Field {
 
   /// Field.textArea is a constructor that helps in managing textArea fields
   Field.textArea(this.key,
-      {this.value,
+      {this.label,
+      this.value,
       this.validate,
       this.autofocus = false,
       this.dummyData,
@@ -685,6 +751,7 @@ class Field {
   /// Field.phoneNumber is a constructor that helps in managing phoneNumber fields
   Field.phoneNumber(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -722,6 +789,7 @@ class Field {
   /// Field.checkbox is a constructor that helps in managing textArea fields
   Field.checkbox(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -811,6 +879,7 @@ class Field {
   /// Field.switchBox is a constructor that helps in managing switch fields
   Field.switchBox(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -900,6 +969,7 @@ class Field {
   /// Field.datetime is a constructor that helps in managing datetime fields
   Field.datetime(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -964,6 +1034,7 @@ class Field {
   /// Field.date is a constructor that helps in managing date fields
   Field.date(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -1028,6 +1099,7 @@ class Field {
   /// Field.chips is a constructor that helps in managing chips fields
   Field.chips(
     this.key, {
+    this.label,
     this.value,
     this.validate,
     this.autofocus = false,
@@ -1090,7 +1162,11 @@ class Field {
   }
 
   /// The key of the field
-  String key;
+  late String key;
+
+  /// The label of the field
+  /// If the label is not provided, the key will be used as the label
+  String? label;
 
   /// The value of the field
   dynamic value;
@@ -1111,7 +1187,7 @@ class Field {
   dynamic style;
 
   /// Get the name of the field
-  String get name => key;
+  String get name => label?.tr() ?? key.titleCase.tr();
 
   /// Get the header of the field
   Widget? header;